requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
httpx==0.28.1
python-dotenv==1.1.1
pydantic==2.11.7
